move plan.py to reinforce
